# server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "SDET School MERN store",
    "description": "SWAGGER API Documentation for Students of SDET School"
  },
  "servers": [
    {
      "url": "https://mern-ecommerce.sdet.school/api"
    },
    {
      "url": "http://localhost:4005/api"
    }
  ],
  "paths": {
...
    "/brand/add": {
      "post": {
        "tags": [
          "Brand"
        ],
        "summary": "Add new brand",
        "description": "Add new brand",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBrand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Add new brand",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddBrandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/brand": {
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Brands by merchant",
        "description": "Brands by merchant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Brand by merchant",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "brands": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BrandMerchantResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/brand/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "description": "Brand id",
          "required": true
        }
      ],
      "get": {
        "tags": [
          "Brand"
        ],
        "summary": "Get Brand",
        "description": "Returns brands products",
        "responses": {
          "200": {
            "description": "Brand's products",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductResponse"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pages": {
                      "type": "integer"
                    },
                    "totalProducts": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Brand not found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Brand"
        ],
        "summary": "Update brand",
        "description": "Update brand",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Brand updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart/add": {
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Create cart",
        "description": "Create cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartCreateUpdate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cart Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/add": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Add order",
        "description": "Add order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrder"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Add order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/order/search": {
      "parameters": [
        {
          "in": "query",
          "name": "search",
          "required": true,
          "description": "search param"
        }
      ],
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Add order",
        "description": "Add order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Order search result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/order": {
      "parameters": [
        {
          "in": "query",
          "name": "page",
          "description": "Page number",
          "required": true,
          "example": 1
        },
        {
          "in": "query",
          "name": "limit",
          "description": "number or orders on the page",
          "required": true,
          "example": 20
        }
      ],
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Add order",
        "description": "Add order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/order/me": {
      "parameters": [
        {
          "in": "query",
          "name": "page",
          "description": "Page number",
          "required": true,
          "example": 1
        },
        {
          "in": "query",
          "name": "limit",
          "description": "number or orders on the page",
          "required": true,
          "example": 20
        }
      ],
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Get user orders",
        "description": "Get user orders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/order/{orderId}": {
      "parameters": [
        {
          "in": "path",
          "description": "Order id",
          "name": "orderId",
          "required": true
        }
      ],
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Get single order",
        "description": "Get single order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Single order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/order/cancel/{orderId}": {
      "parameters": [
        {
          "in": "path",
          "name": "orderId",
          "description": "Order id",
          "required": true
        }
      ],
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Delete single order",
        "description": "Delete single order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cancel order",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/order/status/item/{itemId}": {
      "parameters": [
        {
          "description": "Product id",
          "name": "itemId",
          "required": true,
          "in": "path"
        }
      ],
      "put": {
        "tags": [
          "Order"
        ],
        "summary": "Cancel item in order",
        "description": "Cancel item in order by product id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderItemCancel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order item cancel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Unauthorized": {
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
      "Error": {
        "properties": {
          "error": {
            "type": "string",
            "example": "Error explanation"
          }
        }
      },
...
      "Success": {
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "AddBrand": {
        "type": "object",
        "required": [
          "name",
          "description",
          "isActive"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "AddBrandResponse": {
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "brand": {
            "properties": {
              "isActive": {
                "type": "boolean"
              },
              "merchant": {
                "type": "string"
              },
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "created": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "__v": {
                "type": "integer"
              }
            }
          }
        }
      },
      "ProductResponse": {
        "properties": {
          "taxable": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "brand": {
            "properties": {
              "name": {
                "type": "string"
              },
              "_id": {
                "type": "string"
              },
              "isActive": {
                "type": "boolean"
              }
            }
          },
          "_id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "price": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageKey": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "example": "2022-11-30T01:25:25.197Z"
          },
          "slug": {
            "type": "string"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
...
      "BrandMerchantResponse": {
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "merchant": {
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
      "BrandUpdate": {
        "properties": {
          "brand": {
            "properties": {
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "description": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "CartCreateUpdate": {
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "properties": {
                "quantity": {
                  "type": "integer",
                  "example": 3
                },
                "price": {
                  "type": "number",
                  "example": 5.43
                },
                "taxable": {
                  "type": "boolean",
                  "example": true
                },
                "product": {
                  "type": "string",
                  "description": "Product id"
                }
              }
            }
          }
        }
      },
      "CreateCartResponse": {
        "properties": {
          "success": {
            "type": "boolean"
          },
          "cartId": {
            "type": "string"
          }
        }
      },
...
      "AddOrder": {
        "properties": {
          "cartId": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        }
      },
      "AddOrderResponse": {
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "order": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "OrderSearchResponse": {
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "properties": {
                "_id": {
                  "type": "string"
                },
                "total": {
                  "type": "number"
                },
                "created": {
                  "type": "string"
                },
                "products": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "purchasePrice": {
                        "type": "number"
                      },
                      "totalPrice": {
                        "type": "number"
                      },
                      "priceWithTax": {
                        "type": "number"
                      },
                      "totalTax": {
                        "type": "number"
                      },
                      "status": {
                        "type": "string"
                      },
                      "_id": {
                        "type": "string"
                      },
                      "quantity": {
                        "type": "integer"
                      },
                      "product": {
                        "type": "object",
                        "properties": {
                          "taxable": {
                            "type": "boolean"
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "brand": {
                            "type": "object",
                            "properties": {
                              "isActive": {
                                "type": "boolean"
                              },
                              "merchant": {
                                "type": "string"
                              },
                              "_id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "created": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              },
                              "__v": {
                                "type": "integer"
                              }
                            }
                          },
                          "_id": {
                            "type": "string"
                          },
                          "sku": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "quantity": {
                            "type": "integer"
                          },
                          "price": {
                            "type": "number"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "imageKey": {
                            "type": "string"
                          },
                          "created": {
                            "type": "string",
                            "example": "2022-11-30T01:25:25.197Z"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "__v": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                },
                "totalTax": {
                  "type": "number"
                },
                "totalWithTax": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
      "OrderItemCancel": {
        "properties": {
          "orderId": {
            "type": "string"
          },
          "cartId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}